+++ draw.js
@@ -6,12 +6,9 @@ function drawPlayer() {
   if (player.maxJump === Infinity) ratio = 1;
   if (player.maxJump === 0) ratio = 0;
   let dRatio = dashTimer / dashDuration;
-  playerDisp.tint = PIXI.utils.rgb2hex([
-    (1 - ratio) * (1 - dRatio),
-    dRatio,
-    ratio * (1 - dRatio)
-  ]);
-  if (editor?.invincible) playerDisp.tint = PIXI.utils.rgb2hex([1, 0, 1]);
+  let tint = [(1 - ratio) * (1 - dRatio), dRatio, ratio * (1 - dRatio)];
+  if (editor?.invincible) tint = [1, 0, 1];
+  playerDisp.tint = PIXI.utils.rgb2hex(tint);
   playerDisp.alpha = player.isDead ? 0.5 : 1;
   playerDisp.x = player.x;
   playerDisp.y = player.y;