+++ optionMenu.js
@@ -1,0 +1,9 @@
+const option = new Vue({
+  el: "#options",
+  data: {
+    optionsOpen: false,
+    options: options,
+    addControl: false
+  }
+});
+if (options.theme === "default") unhide();